@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "veg2hab"
-version = "1.0.0"
+version = "1.1.0a0"
 description = "Package voor automatisch omzetten van vegetatiekarteringen naar habitatkaarten"
 authors = ["Spheer.ai <info@spheer.ai>"]
 readme = "README.md"
@@ -9,9 +9,9 @@ classifiers = [
   "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
 
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.7",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
 ]
 
 [project.urls]
@@ -23,25 +23,25 @@ Source = "https://github.com/Spheer-ai/veg2hab"
 veg2hab = "veg2hab.__main__:veg2hab"
 
 [tool.poetry.dependencies]
-python = ">=3.7.1,<4.0"
-pandas = "^1.3.0"
-geopandas = "^0.10.0"
-openpyxl = "^3.1.0"
+python = ">=3.10,<4.0"
+pandas = "^2.2.3"
+geopandas = "^1.0.1"
+openpyxl = "^3.1.5"
 xlrd = "^2.0.1"
-pyodbc= "^4.0.1"
-pydantic = "^1.10.0"
-setuptools = "<60.0.0"
-pygeos = "^0.14"
-typing-extensions = "4.7.1"
-click = "^8.1.7"
-fiona = ">=1.8,<1.10"
+pyodbc = "^5.2.0"
+pydantic = "^2.10.6"
+setuptools = "^75.8.0"
+click = "^8.1.8"
+fiona = "^1.10.1"
+shapely = "^2.0.7"
+pydantic-settings = "^2.8.0"
 
 [tool.poetry.group.dev.dependencies]
-ipykernel = ">=6.0.0"
-black = ">=23.0.0"
-isort = ">=5.0.0"
-pytest = ">=7.0.0"
-pyright = "^1.1.359"
+ipykernel = "^6.29.5"
+black = "^25.1.0"
+isort = "^6.0.0"
+pytest = "^8.3.4"
+pyright = "^1.1.394"
 
 [build-system]
 requires = ["poetry-core"]
