@@ -204,8 +204,8 @@ def test_splittend():
                 2,
             ],
             "geometry": [
-                Polygon([(2, 1), (2, 2), (3, 2), (3, 1)]),
                 Polygon([(0, 1), (0, 2), (1, 2), (1, 1)]),
+                Polygon([(2, 1), (2, 2), (3, 2), (3, 1)]),
                 Polygon([(1, 0), (1, 3), (2, 3), (2, 0)]),
             ],
         }
