@@ -1,11 +1,5 @@
 import warnings
 
-__version__ = "1.0.0"
-
-# Filter out the following warning message
-warnings.filterwarnings(
-    "ignore",
-    message="The Shapely GEOS version .* is incompatible with the GEOS version PyGEOS was compiled with .*",
-)
+__version__ = "1.1.0a0"
 
 from .main import bronbestanden, installatie_instructies, run
